+++ playground/app/pages/[...slug].vue
@@ -1,34 +1,21 @@
 <script lang="ts" setup>
-import { queryContent, useRoute } from '#imports'
-import { useAsyncData } from '#app'
+import { queryCollection, useAsyncData, useRoute } from '#imports'
 
 const route = useRoute()
-const { data, error } = await useAsyncData(route.path, () => queryContent(route.path ?? '/').findOne())
+const path = route.path.replace(/\/$/, '') || '/'
+const { data, error } = await useAsyncData(path, () => queryCollection('content').path(path).first())
 </script>
 
 <template>
   <main>
     <article v-if="data">
-      <!-- 'list' transformer content -->
-      <template v-if="data._extension === 'list'">
-        <p>This is a custom "list" transformer:</p>
-        <ul>
-          <li v-for="item in data.body" :key="item">
-            {{ item }}
-          </li>
-        </ul>
-      </template>
-
-      <!-- any other markdown content -->
-      <article v-else>
-        <ContentRenderer :value="data" />
-      </article>
+      <ContentRenderer :value="data" />
     </article>
 
     <!-- 404 not found -->
-    <template v-else-if="error">
+    <template v-else>
       <p>No content found for "{{ route.path }}"</p>
-      <pre>{{ error.data }}</pre>
+      <pre v-if="error">{{ error }}</pre>
     </template>
   </main>
 </template>